requirements.txt
python-redmine==1.5.1
PyYAML==3.11
requests==2.10.0
